% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xCell.R
\name{v_chmFoldChangeLog10_xCell}
\alias{v_chmFoldChangeLog10_xCell}
\title{Calculate log10 fold-change value for module: xCell}
\usage{
v_chmFoldChangeLog10_xCell(
  outputFolderPath,
  log10Threshold,
  grpName_fc,
  filterNoTPM
)
}
\description{
(Internal) Helper function, used to calculate log10 fold-change value for module xCell, should be called within the main function (v_chmXcell).
}
\keyword{internal}
