% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CHM_G.R
\name{v_chmRiskScore}
\alias{v_chmRiskScore}
\title{Calculate risk score for module: CHM_G (Gene)}
\usage{
v_chmRiskScore()
}
\description{
(Internal) Helper function, used to calcualte risk score for module CHM_G (Gene), should be called within the main function (v_chmSignaturePanel). Please note that currently the risk score calculation method is under revision and in preparation for a peer-reviewed publication. In order to keep up with the standard of vigilante, risk score calculation section in module CHM_G will be temporarily deactivated and will be reactivated once the related publication goes alive.
}
\keyword{internal}
