% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CHM_G.R
\name{v_chmTPM2RHKG}
\alias{v_chmTPM2RHKG}
\title{Convert transcripts per million (TPM) to ratio to housekeeping genes (RHKG) for module: CHM_G (Gene)}
\usage{
v_chmTPM2RHKG()
}
\description{
(Internal) Helper function, experimental, used to convert transcripts per million (TPM) to ratio to housekeeping genes (RHKG) for module: CHM_G (Gene), should be called within the main function (v_chmSignaturePanel). The goal behind this method is to provide an altertive measurement of gene expression that can help reduce batch effects and system variance, and make samples across studies more comparable.
}
\keyword{internal}
