% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CHM_G.R
\name{v_magick_chm}
\alias{v_magick_chm}
\title{Combine the main body and side annotation of heatmap for module: CHM_G/T (both Gene and Transcript)}
\usage{
v_magick_chm(outputFolderPath, signaturePanelType, sigType, moduleType)
}
\description{
(Internal) Helper function, used to combine the main body and side annotation of heatmap for module CHM_G/T (both Gene and Transcript), should be called within the respective main function (v_chmSignaturePanel for Gene and v_chmTranscript for Transcript).
}
\keyword{internal}
