% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xCell.R
\name{v_magick_xCell}
\alias{v_magick_xCell}
\title{Combine the main body and side annotation of heatmap for module xCell}
\usage{
v_magick_xCell(outputFolderPath, sigType)
}
\description{
(Internal) Helper function, used to combine the main body and side annotation of heatmap for module xCell, should be called within the main function (v_chmXcell).
}
\keyword{internal}
