% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Circos.R
\name{v_prepareVdata_Circos}
\alias{v_prepareVdata_Circos}
\title{Prepare data for module: Circos}
\usage{
v_prepareVdata_Circos(
  doBAF = FALSE,
  doCNA = FALSE,
  doMB = c(FALSE, "somatic", "germline"),
  doGE = c(FALSE, "log"),
  doFL = FALSE,
  unifyTrackYlim = TRUE
)
}
\arguments{
\item{doBAF}{logic, whether to prepare B Allele Frequency (BAF) data, if FALSE, no BAF data will be available for downstream v_Circos function and no BAF track will be shown on Circos plots; if TRUE, please make sure BAF data files are properly named, see Details for more information about file naming.}

\item{doCNA}{logic, whether to prepare Copy Number Alteration (CNA) data, if FALSE, no CNA data will be available for downstream v_Circos function and no CNA track will be shown on Circos plots; if TRUE, please make sure CNA data files are properly named, see Details for more information about file naming.}

\item{doMB}{mixed vector, c(FALSE, "somatic", "germline"), whether to prepare Mutation Burden (MB) data, if FALSE, no MB data will be available for downstream v_Circos function and no MB track will be shown on Circos plots; if TRUE, please follow the instruction and run v_llpMultiLoader function first. Also, if TRUE, specify the mutation type, can be either "somatic" or "germline" or both, each type will be shown as a separate track on Circos plots.}

\item{doGE}{mixed vector, c(FALSE, "log"), whether to prepare Gene Expression (GE) data, if FALSE, no GE data will be available for downstream v_Circos function and no GE track will be shown on Circos plots; if TRUE, please make sure GE data files are properly named, see Details for more information about file naming. Also, if TRUE, specify the value format, can be either "log" or "TPM", choose "TPM" for showing raw TPM value, choose "log" for showing value after log10 conversion (0 will be converted into negative value based on the value range across all samples).}

\item{doFL}{logic, whether to prepare Fusion Link (FL) data, if FALSE, no FL data will be available for downstream v_Circos function and no FL track will be shown on Circos plots; if TRUE, please make sure FL data files are properly named, see Details for more information about file naming.}

\item{unifyTrackYlim}{logic, whether to unify each track's y-axis range based on the value range across all samples, default TRUE.}
}
\value{
list, because R CMD check discourages assignments to the global environment within functions, user needs to run the function with explicitly assigning the return value to a global variable named "prepareVdata_Circos_returnList", which will be a list containing the required variables for downstream analyses.
}
\description{
The workflow of vigilante is highly module-based. To ensure a successful and smooth run, vigilante needs to prepare input data before continuing.
}
\details{
Oftentimes input data files generated by upstream tools came with diverse naming conventions. It might be easy for the user to recognize those files, but not for vigilante if there is no recognizable patterns.

To make input data files clear to vigilante, it would be nice to have them named something like "studyID_sampleID_(other descriptions).file extension". Here "studyID" is the name of the study or project, and it will be used in multiple naming situations (such as on the plot, or in the output file names), so it is recommended to be concise and meaningful.

For module Circos, currently supported input data files are listed below, please contact the author if you want to add more files to the supported list:
B Allele Frequency (BAF): *baf.tsv
Copy Number Alteration (CNA): *cna.tsv
Mutation Burden (MB): see the help file of v_llpMultiLoader function for more information
Gene Expression (GE): *cDNA_genes.sf for Salmon
Fusion Link (FL): *-fusion-genes.txt for FusionCatcher, *star-fusion*abridged* for STAR-Fusion
}
