% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Circos.R
\name{v_Circos}
\alias{v_Circos}
\title{Main function for module: Circos}
\usage{
v_Circos(
  outputFolderPath = "./_Viz/_circos/",
  startNum = NULL,
  endNum = NULL,
  doBAF = FALSE,
  doCNA = FALSE,
  doMB = c(FALSE, "somatic", "germline"),
  doGE = c(FALSE, "log"),
  doFL = FALSE
)
}
\arguments{
\item{outputFolderPath}{string, relative or absolute path to the output folder, can be set to NULL (no output file will be written to the file system), default "./_Viz/_circos/".}

\item{startNum}{integer, by default, vigilante will go through every individual sample and sample group, generating a Circos plot per each item in the list. If an integer is specified, vigilante will instead start generating Circos plots at the specified index (e.g. there 10 samples and 2 groups in the list, set startNum to 11 will make vigilante start generating Circos plots at the first group). If set to NULL, vigilante will automatically detect Circos pltos already existing in the output folder and continue at the most recent generated one (e.g. there are 12 Circos plots in total to be generated and 6 of them already exist in the output folder, to make sure all plots are fully-rendered, vigilante will continue at the 6th plot and overwrite it). Since generating comprehensive Circos plots is very computational-intensive, it is possible that function run get interrupted and half-rendered plots get generated because of running out of memory. Therefore, it is recommended to set startNum to NULL so that vigilante can ensure every item in the list to be plotted get fully-rendered.}

\item{endNum}{integer, by default, vigilante will go through every individual sample and sample group, generating a Circos plot per each item in the list. If an integer is specified, vigilante will instead stop generating Circos plots at the specified index (e.g. there 10 samples and 2 groups in the list, set endNum to 6 will make the 6th sample as the last one to be plotted, and there will be no Circos plots for items after the 6th sample).}

\item{doBAF}{logic, whether to show BAF track on Circos plots.}

\item{doCNA}{logic, whether to show CNA track on Circos plots.}

\item{doMB}{mixed vector, c(FALSE, "somatic", "germline"), whether to show MB track on Circos plots; if TRUE, specify the mutation type, can be either "somatic" or "germline" or both, each type will be shown as a separate track on Circos plots.}

\item{doGE}{mixed vector, c(FALSE, "log"), whether to show GE track on Circos plots; if TRUE, specify the value format, can be either "log" or "TPM", choose "TPM" for showing raw TPM value, choose "log" for showing value after log10 conversion (0 will be converted into negative value based on the value range across all samples), must match the value format chosen in v_prepareVdata_Circos function.}

\item{doFL}{logic, whether to show FL track will be shown on Circos plots.}
}
\value{
NULL, when a valid outputFolderPath is provided, output Circos plots will be generated and saved in the provided location, otherwise function run will stop and nothing will be written into the file system.
}
\description{
Extract, clean and reform NGS data, and then generate comprehensive Circos plot with each track (circle) representing different types of data. Currently supported tracks: BAF (B allele frequency), CNA (copy number alterations), mutation burden (somatic/germline mutations), gene expression (TPM) and gene fusions (cross-check validation is available when 2 or more fusion callers are in use; also, fusion status check and validation through statistics are available via module CSE).
}
\details{
Oftentimes input data files generated by upstream tools came with diverse naming conventions. It might be easy for the user to recognize those files, but not for vigilante if there is no recognizable patterns.

To make input data files clear to vigilante, it would be nice to have them named something like "studyID_sampleID_(other descriptions).file extension". Here "studyID" is the name of the study or project, and it will be used in multiple naming situations (such as on the plot, or in the output file names), so it is recommended to be concise and meaningful.

For module Circos, currently supported input data files are listed below, please contact the author if you want to add more files to the supported list:
B Allele Frequency (BAF): *baf.tsv
Copy Number Alteration (CNA): *cna.tsv
Mutation Burden (MB): see the help file of v_llpMultiLoader function for more information
Gene Expression (GE): *cDNA_genes.sf for Salmon
Fusion Link (FL): *-fusion-genes.txt for FusionCatcher, *star-fusion*abridged* for STAR-Fusion
}
